Data Cleaning - Update 3
I) Checked for consistent values for features. Found two affected rows with wrong permutation of columns.

II) Removed duplicate rows based on all features but 'order'.

Added summary of the data cleaning process at the end of the notebook.

Going to continue feature exploration in Tableau and a new notebook.